// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-header-deliverable-model.ts
export class RFQSupplierHeaderDeliverableModel {
    rfqHeaderDeliverableId: string;
    rfqId: string;
    supplierId: string;
    milestone: string;
    milestoneName: string;
    description: string;
    previousMilestone: string;
    progress: string;
    stageProgress: string;
    beginDate: string;
    beginDateISO: string;
    surrogateById: string;
    surrogationDate: string;
    comments: string;
}