// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-header-note-model.ts
export class RFQSupplierHeaderNoteModel {
    id: string;
    description: string;
    buyDescription: string;
    rfqId: string;
    supplierId: string;
    rfqHeaderNoteId: string;
    surrogateById: string;
    surrogationDate: string;
    attributeCategoryId: string;
    comments: string;
    supplierResponse: string;
}